SymbolDatabase: flatten Configuration module into single file
Inline `SymbolDatabase::Configuration::Settings` into
`lib/datadog/symbol_database/configuration.rb` and remove the empty
wrapper plus `configuration/settings.rb`. Mirror the RBS layout in
`sig/`. Update the two require sites
(`core/configuration/settings.rb` and the spec).

Matches the single-file pattern already used by
`lib/datadog/open_feature/configuration.rb`. No behavior change.
